playwright-demos/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.5",
    "pytest-playwright>=0.7.0",
]

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 120

[tool.bandit]
exclude_dirs = ["tests", "path/to/file"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[dependency-groups]
dev = []
